Updated Video and User classes have integer instead of long
Fixed User issue with table.

+++ src/main/java/com/edufy/edufy/services/UserServiceInterface.java
@@ -9,11 +9,11 @@ import java.util.List;
 public interface UserServiceInterface {
     List<User> getAllUsers();
 
-    User getUser(User user, long id);
+    User getUser(User user, int id);
 
     User addUser(User user);
 
-    User updateUser(User user, long id);
+    User updateUser(User user, int id);
 
-    void removeUser(long id);
+    void removeUser(int id);
 }
